Fix size of transient window. Relocate stack.
Relocate stack a bit higher in memory in the hopes that will
gain us some room (it's also a rounder number). Also move the
palette to just after the vehicule stats. This means in theory
the palette can smash the stack, but it should not do so before
at least 365 deep on stack (well you have to count registers,
but that's still rather deep especially since there's view movem
calls in that code)