Center APPUYEZ SUR START properly.
Actually that forced me to change the copyright art
slightly and tighten it up, so it ends up one character
smaller. Given that the ampersand already took two cells
(one for a liiitle bit at the left, 1 pel wide, the other
8 pels for the rest) I just moved everything by 4 pels
over to end up with a half space between TM and (C). I
think this looks better actually anyway, and as a bonus
everything is centered.