Fix subtitle.
This took some doing because the original source converted to
the wrong palette, which isn't surprising because I was too lazy
to do it properly. So there are a couple of Python scripts that
do the work of rewriting the bitmap with the right pixel values
and another to optimize the tile map. It's not 100% great but
it's serviceable.