WIP: improve text output for descriptions.
Still not quite there yet, size calculations
are complicated by the fact that in one case we
want to output n chars and in the other case all we're
interested in is where's the next output cell.

d1 for char-by-char output has a totally different
requirement from d1 for whole-line-at-the-time output.

Consider modifying routine to use a1 and ignore d1 altogether,
not allow flow of characters to occur at all. Split in
separate routine for dialogue text when d4 is set to 1.
Probably disable whole mechanism checking for out of bounds
characters in d4 = 1 case (and disable out of bounds checks
anyway, since in prop font they won't work righ)