Add laser sword "easter egg".
The long description of Laser Sword in French is a bit of
a Star Wars easter egg, but nobody could see it since it's
not sold in shops. So I added it to Meese. It doesn't change
play balance there, so it's purely "fan service". Wonder who
will catch that :)