Optimize text output routines.
This is a very nice (if I do say so myself) hand-optimized
inner loop which was replaced with a 16-entry lookup table (well,
two of them really) and looking up instructions timings in the
Neo Geo dev wiki.

So basically instead of having 16 * 8 iterations per character cell
(which is even worse when the character has two cells!) we now have
have 16 iterations per character cell, with the 8 iteration inner
loop replaced by a bunch of table lookups.

This was done to save time, but mostly space, in preparation of
moving those routines back to the original location.

I still have to optimize them slightly, mostly the two separate
"non-title" routines (that's the routine that shows up towards
the end game; maybe I can just have that one located elsewhere;
the instructions I saved removing the loop should be plenty to add
a conditional jump at the start of the routine; but ideally the
routine should be written in such a way that it can be used in
both modes, the only reason it can't is because of the way the
d1/d2 registers are used to keep track of the current cell in
each mode, it's considerably different although in fixed width
you get the exact same result). But this should give us a fighting
chance since now the inner loop is much shorter, this frees up
some of the space we had to add because of the "second cell"
calculation when a character straddles the cell boundary.