Further optimization of character output.
We now use register d2 internally to flag whether we need
to subtract one from the d1 register prior to rts.

This is mostly done to save space and reuse a maximum of code,
in prevision of moving this routine in the "empty spaces" left by
removal of the old obsolete routines.