Rearrange code to get smallest possible patch.
This is the best I can do without doing crazy stuff (like
changing the dc.b $FF while macro to exact db directives
for existing stuff. But this starts getting pretty crazy and
I don't want to go there as it's incredibly brittle if we
add any kind of bugfixes.