Finalize the dialogue output routine, consolidated.
Now I think it's as small as possible. it's actually smaller
than the legacy regular output version (and the atonce output
version was smaller because it didn't have to monkey with some
extra flags the incremental one had to play with)

Only thing left I'd like to fix is this:

sub.b	d0, d3
neg.b	d3
addq.b	#8, d3
sub.b	d3, d0

There *has* to be a better instruction sequence for this.
Note that d0 and d3 have to have the exact values that
end up there right now, but I can't escape the feeling that
I could get rid of that neg routine with some clever subtraction.